requirements.txt
# Core dependencies
openai>=1.12.0
python-dotenv>=1.0.0
rich>=13.7.0
pydantic>=2.5.0

# Development dependencies
pytest>=7.4.3
pytest-asyncio>=0.23.3
black>=23.12.1
ruff>=0.1.9
